refactor: deduplicate verifier test helpers
- Move duplicated helper functions to shared helpers.rs module
- Remove duplicate setup_mock_server_with_solver_registry functions
- Remove duplicate create_solver_registry_resource functions
- Remove duplicate build_test_config_with_mock_server functions
- Update all test files to use shared helpers

Tests pass: Move 62, EVM 48, Verifier (Rust) 178, Solver (Rust) 119
