+++ trusted-verifier/tests/cross_chain_tests.rs
@@ -3,94 +3,10 @@
 //! These tests verify that escrow events can be matched to intent events
 //! across different chains using intent_id.
 
-use serde_json::json;
-use trusted_verifier::config::Config;
 use trusted_verifier::monitor::{EscrowEvent, IntentEvent};
-use trusted_verifier::validator::CrossChainValidator;
-use wiremock::matchers::{method, path};
-use wiremock::{Mock, MockServer, ResponseTemplate};
 #[path = "mod.rs"]
 mod test_helpers;
-use test_helpers::{build_test_config_with_mvm, create_base_escrow_event, create_base_intent_mvm};
-
-// ============================================================================
-// HELPER FUNCTIONS
-// ============================================================================
-
-/// Build a test config with a mock server URL
-fn build_test_config_with_mock_server(mock_server_url: &str) -> Config {
-    let mut config = build_test_config_with_mvm();
-    config.hub_chain.rpc_url = mock_server_url.to_string();
-    config
-}
-
-/// Helper to create a mock SolverRegistry resource response with MVM address
-fn create_solver_registry_resource_with_mvm_address(
-    registry_address: &str,
-    solver_address: &str,
-    connected_chain_mvm_address: Option<&str>,
-) -> serde_json::Value {
-    let solver_entry = if let Some(mvm_addr) = connected_chain_mvm_address {
-        json!({
-            "key": solver_address,
-            "value": {
-                "public_key": [1, 2, 3, 4],
-                "connected_chain_evm_address": {"vec": []},
-                "connected_chain_mvm_address": {"vec": [mvm_addr]},
-                "registered_at": 1234567890
-            }
-        })
-    } else {
-        json!({
-            "key": solver_address,
-            "value": {
-                "public_key": [1, 2, 3, 4],
-                "connected_chain_evm_address": {"vec": []},
-                "connected_chain_mvm_address": {"vec": []},
-                "registered_at": 1234567890
-            }
-        })
-    };
-
-    json!([{
-        "type": format!("{}::solver_registry::SolverRegistry", registry_address),
-        "data": {
-            "solvers": {
-                "data": [solver_entry]
-            }
-        }
-    }])
-}
-
-/// Setup a mock server with solver registry for MVM tests
-async fn setup_mock_server_with_solver_registry(
-    solver_address: Option<&str>,
-    connected_chain_mvm_address: Option<&str>,
-) -> (MockServer, CrossChainValidator) {
-    let mock_server = MockServer::start().await;
-    let registry_address = "0x1";
-
-    if let Some(solver_addr) = solver_address {
-        let resources_response = create_solver_registry_resource_with_mvm_address(
-            registry_address,
-            solver_addr,
-            connected_chain_mvm_address,
-        );
-
-        Mock::given(method("GET"))
-            .and(path(format!("/v1/accounts/{}/resources", registry_address)))
-            .respond_with(ResponseTemplate::new(200).set_body_json(resources_response))
-            .mount(&mock_server)
-            .await;
-    }
-
-    let config = build_test_config_with_mock_server(&mock_server.uri());
-    let validator = CrossChainValidator::new(&config)
-        .await
-        .expect("Failed to create validator");
-
-    (mock_server, validator)
-}
+use test_helpers::{create_base_escrow_event, create_base_intent_mvm, setup_mock_server_with_solver_registry};
 
 // ============================================================================
 // TESTS
