+++ trusted-verifier/tests/helpers.rs
@@ -5,11 +5,14 @@
 use base64::{engine::general_purpose, Engine as _};
 use ed25519_dalek::SigningKey;
 use rand::{Rng, RngCore};
+use serde_json::json;
 use trusted_verifier::config::{ApiConfig, ChainConfig, Config, EvmChainConfig, VerifierConfig};
 use trusted_verifier::evm_client::EvmTransaction;
 use trusted_verifier::monitor::{ChainType, EscrowEvent, FulfillmentEvent, IntentEvent};
 use trusted_verifier::mvm_client::MvmTransaction;
-use trusted_verifier::validator::FulfillmentTransactionParams;
+use trusted_verifier::validator::{CrossChainValidator, FulfillmentTransactionParams};
+use wiremock::matchers::{method, path};
+use wiremock::{Mock, MockServer, ResponseTemplate};
 
 /// Build a valid in-memory test configuration with a fresh Ed25519 keypair.
 /// Keys are encoded using standard Base64 and set as environment variables.
@@ -303,4 +306,225 @@ pub fn create_base_evm_transaction() -> EvmTransaction {
         gas_price: "0x3b9aca00".to_string(), // 1 Gwei (1,000,000,000 wei) - typical test value
         status: Some("0x1".to_string()), // Success
     }
+}
+
+/// Build a test config with a mock server URL
+#[allow(dead_code)]
+pub fn build_test_config_with_mock_server(mock_server_url: &str) -> Config {
+    let mut config = build_test_config_with_mvm();
+    config.hub_chain.rpc_url = mock_server_url.to_string();
+    config
+}
+
+/// Helper to create a mock SolverRegistry resource response with MVM address
+/// SimpleMap<address, SolverInfo> is serialized as {"data": [{"key": address, "value": SolverInfo}, ...]}
+#[allow(dead_code)]
+pub fn create_solver_registry_resource_with_mvm_address(
+    registry_address: &str,
+    solver_address: &str,
+    connected_chain_mvm_address: Option<&str>,
+) -> serde_json::Value {
+    let solver_entry = if let Some(mvm_addr) = connected_chain_mvm_address {
+        json!({
+            "key": solver_address,
+            "value": {
+                "public_key": [1, 2, 3, 4],
+                "connected_chain_evm_address": {"vec": []},
+                "connected_chain_mvm_address": {"vec": [mvm_addr]},
+                "registered_at": 1234567890
+            }
+        })
+    } else {
+        json!({
+            "key": solver_address,
+            "value": {
+                "public_key": [1, 2, 3, 4],
+                "connected_chain_evm_address": {"vec": []},
+                "connected_chain_mvm_address": {"vec": []},
+                "registered_at": 1234567890
+            }
+        })
+    };
+
+    json!([{
+        "type": format!("{}::solver_registry::SolverRegistry", registry_address),
+        "data": {
+            "solvers": {
+                "data": [solver_entry]
+            }
+        }
+    }])
+}
+
+/// Helper to create a mock SolverRegistry resource response with EVM address
+/// SimpleMap<address, SolverInfo> is serialized as {"data": [{"key": address, "value": SolverInfo}, ...]}
+#[allow(dead_code)]
+pub fn create_solver_registry_resource_with_evm_address(
+    registry_address: &str,
+    solver_address: &str,
+    evm_address: Option<&str>,
+) -> serde_json::Value {
+    let solver_entry = if let Some(evm_addr) = evm_address {
+        // Convert hex string (with or without 0x) to vector<u8>
+        let addr_clean = evm_addr.strip_prefix("0x").unwrap_or(evm_addr);
+        let bytes: Vec<u64> = (0..addr_clean.len())
+            .step_by(2)
+            .map(|i| u8::from_str_radix(&addr_clean[i..i + 2], 16).unwrap() as u64)
+            .collect();
+
+        // SolverInfo with connected_chain_evm_address set
+        json!({
+            "key": solver_address,
+            "value": {
+                "public_key": [1, 2, 3, 4], // Dummy public key bytes
+                "connected_chain_evm_address": {"vec": [bytes]}, // Some(vector<u8>)
+                "connected_chain_mvm_address": {"vec": []}, // None
+                "registered_at": 1234567890
+            }
+        })
+    } else {
+        // SolverInfo without connected_chain_evm_address
+        json!({
+            "key": solver_address,
+            "value": {
+                "public_key": [1, 2, 3, 4], // Dummy public key bytes
+                "connected_chain_evm_address": {"vec": []}, // None
+                "connected_chain_mvm_address": {"vec": []}, // None
+                "registered_at": 1234567890
+            }
+        })
+    };
+
+    json!([{
+        "type": format!("{}::solver_registry::SolverRegistry", registry_address),
+        "data": {
+            "solvers": {
+                "data": [solver_entry]
+            }
+        }
+    }])
+}
+
+/// Setup a mock server with solver registry for MVM tests
+/// Returns the mock server and CrossChainValidator
+#[allow(dead_code)]
+pub async fn setup_mock_server_with_solver_registry(
+    solver_address: Option<&str>,
+    connected_chain_mvm_address: Option<&str>,
+) -> (MockServer, CrossChainValidator) {
+    let mock_server = MockServer::start().await;
+    let registry_address = "0x1";
+
+    if let Some(solver_addr) = solver_address {
+        let resources_response = create_solver_registry_resource_with_mvm_address(
+            registry_address,
+            solver_addr,
+            connected_chain_mvm_address,
+        );
+
+        Mock::given(method("GET"))
+            .and(path(format!("/v1/accounts/{}/resources", registry_address)))
+            .respond_with(ResponseTemplate::new(200).set_body_json(resources_response))
+            .mount(&mock_server)
+            .await;
+    }
+
+    let config = build_test_config_with_mock_server(&mock_server.uri());
+    let validator = CrossChainValidator::new(&config)
+        .await
+        .expect("Failed to create validator");
+
+    (mock_server, validator)
+}
+
+/// Setup a mock server with solver registry for monitor tests
+/// Returns the mock server and Config
+#[allow(dead_code)]
+pub async fn setup_mock_server_with_solver_registry_config(
+    solver_address: Option<&str>,
+    connected_chain_mvm_address: Option<&str>,
+) -> (MockServer, Config) {
+    let mock_server = MockServer::start().await;
+    let registry_address = "0x1";
+
+    if let Some(solver_addr) = solver_address {
+        let resources_response = create_solver_registry_resource_with_mvm_address(
+            registry_address,
+            solver_addr,
+            connected_chain_mvm_address,
+        );
+
+        Mock::given(method("GET"))
+            .and(path(format!("/v1/accounts/{}/resources", registry_address)))
+            .respond_with(ResponseTemplate::new(200).set_body_json(resources_response))
+            .mount(&mock_server)
+            .await;
+    }
+
+    let config = build_test_config_with_mock_server(&mock_server.uri());
+    (mock_server, config)
+}
+
+/// Setup a mock server that responds to get_resources calls with SolverRegistry (MVM)
+/// Returns the mock server and CrossChainValidator
+#[allow(dead_code)]
+pub async fn setup_mock_server_with_registry_mvm(
+    registry_address: &str,
+    solver_address: &str,
+    connected_chain_mvm_address: Option<&str>,
+) -> (MockServer, CrossChainValidator) {
+    let mock_server = MockServer::start().await;
+
+    let resources_response = create_solver_registry_resource_with_mvm_address(
+        registry_address,
+        solver_address,
+        connected_chain_mvm_address,
+    );
+
+    Mock::given(method("GET"))
+        .and(path(format!("/v1/accounts/{}/resources", registry_address)))
+        .respond_with(ResponseTemplate::new(200).set_body_json(resources_response))
+        .mount(&mock_server)
+        .await;
+
+    let mut config = build_test_config_with_mvm();
+    config.hub_chain.rpc_url = mock_server.uri();
+    let validator = CrossChainValidator::new(&config)
+        .await
+        .expect("Failed to create validator");
+
+    (mock_server, validator)
+}
+
+/// Setup a mock server that responds to get_resources calls with SolverRegistry (EVM)
+/// Returns the mock server and CrossChainValidator
+#[allow(dead_code)]
+pub async fn setup_mock_server_with_registry_evm(
+    registry_address: &str,
+    solver_address: &str,
+    evm_address: Option<&str>,
+) -> (MockServer, CrossChainValidator) {
+    let mock_server = MockServer::start().await;
+
+    let resources_response = create_solver_registry_resource_with_evm_address(
+        registry_address,
+        solver_address,
+        evm_address,
+    );
+
+    Mock::given(method("GET"))
+        .and(path(format!("/v1/accounts/{}/resources", registry_address)))
+        .respond_with(ResponseTemplate::new(200).set_body_json(resources_response))
+        .mount(&mock_server)
+        .await;
+
+    let mut config = build_test_config_with_evm();
+    config.hub_chain.rpc_url = mock_server.uri();
+    // Clear MVM chain config so validator uses EVM path
+    config.connected_chain_mvm = None;
+    let validator = CrossChainValidator::new(&config)
+        .await
+        .expect("Failed to create validator");
+
+    (mock_server, validator)
 }
